Excel -  Airbnb Project
The given public dataset is part of Airbnb which includes all needed information to find out more about hosts, geographical availability, necessary metrics to make predictions and draw conclusions. Since 2008, guests and hosts have used Airbnb to expand on traveling possibilities and present more unique, personalized way of experiencing the world. This dataset describes the listing activity and metrics in NYC, NY for 2019. There are 16 features in this dataset.